+++ core-bundle/src/ContaoCoreBundle.php
@@ -17,6 +17,7 @@ use Contao\CoreBundle\DependencyInjection\Compiler\AddSessionBagsPass;
 use Contao\CoreBundle\DependencyInjection\Compiler\DoctrineMigrationsPass;
 use Contao\CoreBundle\DependencyInjection\Compiler\PickerMenuProviderPass;
 use Contao\CoreBundle\DependencyInjection\ContaoCoreExtension;
+use Symfony\Component\Console\Application;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -37,6 +38,14 @@ class ContaoCoreBundle extends Bundle
     public function getContainerExtension()
     {
         return new ContaoCoreExtension();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function registerCommands(Application $application)
+    {
+        // disable automatic command registration
     }
 
     /**