+++ core-bundle/src/DependencyInjection/ContaoCoreExtension.php
@@ -50,7 +50,7 @@ class ContaoCoreExtension extends ConfigurableExtension
 
         return new Configuration(
             $container->getParameter('kernel.debug'),
-            $container->getParameter('kernel.root_dir')
+            $container->getParameter('kernel.project_dir')
         );
     }
 
@@ -68,7 +68,6 @@ class ContaoCoreExtension extends ConfigurableExtension
             $loader->load($file);
         }
 
-        $container->setParameter('contao.root_dir', $mergedConfig['root_dir']);
         $container->setParameter('contao.web_dir', $mergedConfig['web_dir']);
         $container->setParameter('contao.prepend_locale', $mergedConfig['prepend_locale']);
         $container->setParameter('contao.encryption_key', $mergedConfig['encryption_key']);
@@ -104,7 +103,7 @@ class ContaoCoreExtension extends ConfigurableExtension
 
         $container->setParameter(
             'contao.image.target_dir',
-            $mergedConfig['root_dir'].'/'.$mergedConfig['image']['target_path']
+            $container->getParameter('kernel.project_dir').'/'.$mergedConfig['image']['target_path']
         );
 
         @trigger_error('Using the contao.image.target_path parameter has been deprecated and will no longer work in Contao 5. Use the contao.image.target_dir parameter instead.', E_USER_DEPRECATED);