+++ core-bundle/tests/Command/SymlinksCommandTest.php
@@ -56,9 +56,9 @@ class SymlinksCommandTest extends TestCase
     public function testOutput()
     {
         $container = new ContainerBuilder();
-        $container->setParameter('kernel.root_dir', $this->getRootDir().'/app');
         $container->setParameter('kernel.logs_dir', $this->getRootDir().'/var/logs');
-        $container->setParameter('contao.root_dir', $this->getRootDir());
+        $container->setParameter('kernel.project_dir', $this->getRootDir());
+        $container->setParameter('kernel.root_dir', $this->getRootDir().'/app');
         $container->setParameter('contao.upload_path', 'app');
 
         $container->set(