+++ core-bundle/tests/DependencyInjection/Compiler/AddResourcesPathsPassTest.php
@@ -39,8 +39,8 @@ class AddResourcesPathsPassTest extends TestCase
         $pass = new AddResourcesPathsPass();
 
         $container = new ContainerBuilder();
+        $container->setParameter('kernel.project_dir', $this->getRootDir());
         $container->setParameter('kernel.root_dir', $this->getRootDir().'/app');
-        $container->setParameter('contao.root_dir', $this->getRootDir());
 
         $container->setParameter('kernel.bundles', [
             'FrameworkBundle' => 'Symfony\Bundle\FrameworkBundle\FrameworkBundle',