+++ core-bundle/tests/DependencyInjection/ContaoCoreExtensionTest.php
@@ -50,6 +50,7 @@ class ContaoCoreExtensionTest extends TestCase
         $container = new ContainerBuilder(
             new ParameterBag([
                 'kernel.debug' => false,
+                'kernel.project_dir' => $this->getRootDir(),
                 'kernel.root_dir' => $this->getRootDir().'/app',
             ])
         );
@@ -78,6 +79,7 @@ class ContaoCoreExtensionTest extends TestCase
         $container = new ContainerBuilder(
             new ParameterBag([
                 'kernel.debug' => false,
+                'kernel.project_dir' => $this->getRootDir(),
                 'kernel.root_dir' => $this->getRootDir().'/app',
             ])
         );